Moved TextParseConfig
