+++ docs/release_notes.md
@@ -17,6 +17,8 @@ title: Release notes
 * The following classes has moved to package `org.jsapar.bean` in order to make them visible externally when introducing modules: 
   * org.jsapar.parse.bean.BeanMap
   * org.jsapar.parse.bean.Xml2BeanMapBuilder
+* The following classes has moved to package `org.jsapar.text` in order to make them visible externally when introducing modules: 
+  * org.jsapar.parse.text.TextParseConfig
 * Removed dependency to `java.beans` package since that imposed dependency to `java.desktop` module.
 ## 2.0
 ### 2.0.2
