+++ src/main/java/org/jsapar/parse/csv/CsvLineParserFactory.java
@@ -1,6 +1,6 @@
 package org.jsapar.parse.csv;
 
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.schema.CsvSchema;
 import org.jsapar.schema.CsvSchemaLine;
