+++ src/main/java/org/jsapar/parse/csv/CsvLineParserMatcher.java
@@ -1,6 +1,6 @@
 package org.jsapar.parse.csv;
 
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.schema.CsvSchemaCell;
 import org.jsapar.schema.CsvSchemaLine;
