+++ src/main/java/org/jsapar/parse/csv/CsvParser.java
@@ -3,7 +3,7 @@ package org.jsapar.parse.csv;
 import org.jsapar.error.ErrorEventListener;
 import org.jsapar.parse.LineEventListener;
 import org.jsapar.parse.line.ValidationHandler;
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.parse.text.TextSchemaParser;
 import org.jsapar.schema.CsvSchema;
 import org.jsapar.schema.SchemaLine;
