+++ src/main/java/org/jsapar/parse/fixed/FWLineParserMatcher.java
@@ -1,6 +1,6 @@
 package org.jsapar.parse.fixed;
 
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.schema.FixedWidthSchemaCell;
 import org.jsapar.schema.FixedWidthSchemaLine;
