+++ src/main/java/org/jsapar/text/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Text formatting and parsing.
+ * Common public classes for parsing and composing text.
  */
 package org.jsapar.text;
