+++ src/main/java/org/jsapar/utils/text/EnumFormat.java
@@ -1,4 +1,4 @@
-package org.jsapar.text;
+package org.jsapar.utils.text;
 
 import java.text.FieldPosition;
 import java.text.Format;
