+++ src/main/java/org/jsapar/utils/text/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Text formatting and parsing.
+ */
+package org.jsapar.utils.text;
