+++ src/test/java/org/jsapar/TextParserTest.java
@@ -1,6 +1,6 @@
 package org.jsapar;
 
-import org.jsapar.parse.text.TextParseConfig;
+import org.jsapar.text.TextParseConfig;
 import org.jsapar.schema.CsvSchema;
 import org.jsapar.schema.Schema;
 import org.junit.Test;
